refactoring endpoints

+++ exchange-app/src/main/java/com/kolyo/exchange/app/dto/CurrencyConversionRequestDTO.java
@@ -2,13 +2,13 @@ package com.kolyo.exchange.app.dto;
 
 import lombok.Data;
 
-import java.math.BigDecimal;
-
 @Data
 
 public class CurrencyConversionRequestDTO {
 
     private String from;
-    private BigDecimal amount;
+
+    private Double amount;
+
     private String to;
 }
