+++ exchange-app/src/main/java/com/kolyo/exchange/app/dto/ExchangeRateResponseDTO.java
@@ -3,13 +3,13 @@ package com.kolyo.exchange.app.dto;
 import lombok.Builder;
 import lombok.Data;
 
-import java.math.BigDecimal;
-
 @Data
 @Builder
 public class ExchangeRateResponseDTO {
 
     private String fromCurrency;
+
     private String toCurrency;
-    private BigDecimal rates;
+
+    private Double rates;
 }
